Bring back the gpio handler and proper globals :cop:
