+++ src/main.py
@@ -38,9 +38,7 @@ VERSION = "0.0.18"
 
 # Modules
 import document_handler
-import overlay_handler
 import camera_handler
-import menu_handler
 
 ################################################################################
 ##                                    Config                                  ##
@@ -137,6 +135,5 @@ camera = None
 overlay = None
 
 # Begin Camera start-up
-camera, overlay = camera_handler.start_camera(config)
-camera_handler.start_preview(camera, config) # Runs main camera loop
+camera, overlay = camera_handler.start_camera(config) # Runs main camera loop
 camera_handler.stop_camera(camera, overlay, config)
